# docs for user service/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/users/login":{"post":{"tags":["user-controller"],"summary":"logs a user in, public endpoint","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/users/approval/{userId}":{"post":{"tags":["user-controller"],"summary":"approves users, available only to merchants","operationId":"approveUsers","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/users/":{"get":{"tags":["user-controller"],"summary":"Retrieves All users, available to merchants only ","operationId":"findAllUsers","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"post":{"tags":["user-controller"],"summary":"Register users, public endpoint","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/users/{id}/exists":{"get":{"tags":["user-controller"],"summary":"Checks if a user exist, service endpoint","operationId":"doesUserExist","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}}},"components":{"schemas":{"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["password","username"]},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"userId":{"type":"integer","format":"int32"},"role":{"type":"string"},"userName":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"roles":{"type":"string"},"email":{"type":"string"},"approved":{"type":"boolean"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"roles":{"type":"string","pattern":"^Merchant|Employee|Customer"},"email":{"type":"string"},"userName":{"type":"string"},"approved":{"type":"boolean"},"password":{"type":"string"}},"required":["email","name","password","roles","userName"]}}}}